cicd: Jacoco 테스트 제외 패턴추가
- 테스트 커버리지 제외 패턴 추가
