+++ build.gradle
@@ -50,9 +50,13 @@ def jacocoExcludePatterns = [
         "**/*Filter*",
         "**/*Resolver*",
         "**/*Entity*",
+        "**/*Volunteer*",
+        "**/*Location*",
         "**/test/**",
-        "**/resources/**"
+        "**/resources/**",
+        "**/domain/**"
 ]
+
 
 jacocoTestCoverageVerification {
 
@@ -64,21 +68,21 @@ jacocoTestCoverageVerification {
             // 클래스 단위로 룰 체크
             element = 'CLASS'
 
-            // 라인 커버리지를 최소 80% 만족
+
             limit {
                 counter = 'LINE'
                 value = 'COVEREDRATIO'
                 minimum = 0.60
             }
 
-            // 브랜치 커버리지를 최소 80% 만족
+
             limit {
                 counter = 'BRANCH'
                 value = 'COVEREDRATIO'
                 minimum = 0.60
             }
 
-            excludes = jacocoExcludePatterns
+            excludes = jacocoExcludePatterns + QDomains
         }
     }
 }
